feat: add metronome
